BigDecimal (ResultsPrinterSimsCDFMultiple:514) now cast explicitly to double before string formatting.
This appears to resolve issue #24: https://code.google.com/a/eclipselabs.org/p/qmul-genome-convergence-pipeline/issues/detail?id=24
But has not been tested thoroughly.

+++ trunk/trunk/src/uk/ac/qmul/sbcs/evolution/sandbox/ResultsPrinterSimsCDFMultiple.java
@@ -511,7 +511,7 @@ public class ResultsPrinterSimsCDFMultiple{
 									 */
 
 									BigDecimal avgOfSummedDiffs = diffsSummed[t].divide(new BigDecimal(nSites),RoundingMode.HALF_EVEN);
-									String AOSD = String.format("%.16f",avgOfSummedDiffs);
+									String AOSD = String.format("%.16f",avgOfSummedDiffs.doubleValue());
 									System.out.print(
 											"\ttH"+referenceTree+"-H"+testTrees[c]+"\t"
 											+AOSD+"\t"
